+++ src/js/bitcoinjs-extensions.js
@@ -1739,6 +1739,28 @@ libs.bitcoin.networks.elastos = {
     wif: 0xef // TODO set this correctly, same as BTC for now
 };
 
+libs.bitcoin.networks.sugarchain = {
+    messagePrefix: '\x18Sugarchain Signed Message:\n',
+    bip32: {
+        public: 0x0488B21E,
+        private: 0x0488ADE4,
+    },
+    pubKeyHash: 0x3f,
+    scriptHash: 0x7d,
+    wif: 0x80
+};
+
+libs.bitcoin.networks.sugarchaintestnet = {
+    messagePrefix: '\x18Sugarchain Signed Message:\n',
+    bip32: {
+        public: 0x045f1cf6,
+        private: 0x045f18bc,
+    },
+    pubKeyHash: 0x42,
+    scriptHash: 0x80,
+    wif: 0xef
+};
+
 // https://github.com/libs.bitcoinjs-lib/blob/3f6f5ef97a1ee1b8337865209282c0095e22b2e7/src/networks.js
 libs.bitcoin.networks.regtest = {
   messagePrefix: '\x18Bitcoin Signed Message:\n',
@@ -1750,4 +1772,15 @@ libs.bitcoin.networks.regtest = {
   pubKeyHash: 0x6f,
   scriptHash: 0xc4,
   wif: 0xef,
+};
+
+libs.bitcoin.networks.argoneum = {
+  messagePrefix: 'unused',
+  bip32: {
+    public: 0x0488b21e,
+    private: 0x0488ade4
+  },
+  pubKeyHash: 0x32,
+  scriptHash: 0x61,
+  wif: 0xbf
 };